+++ lib/Fakturoid/Exception.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Fakturoid;
+
+class Exception extends \Exception
+{
+}