+++ lib/Fakturoid/Requester.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Fakturoid;
+
+// For testing purposes.
+class Requester
+{
+    public function run($options)
+    {
+        $request  = new Request($options);
+        $response = $request->run();
+
+        return $response;
+    }
+}