+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 - Minimum PHP version was raised from 5.2 to 5.3 due to the need of anonymous
   functions and namespaces.
 
+- Classes were put under `Fakturoid` namespace.
+
+  ```php
+  // Before
+  $f = new Fakturoid('..slug..', '..user@email.cz..', '..api_key..', 'PHPlib <your@email.cz>');
+  try {} catch (FakturoidException $e) {}
+
+  // After
+  $f = new Fakturoid\Client('..slug..', '..user@email.cz..', '..api_key..', 'PHPlib <your@email.cz>');
+  try {} catch (Fakturoid\Exception $e) {}
+  ```
+
 - Method names were changed from `underscored_names` to `camelCase` to be PSR-2 friendly.
 
   ```php